+++ js/util.js
@@ -1,0 +1,11 @@
+import './data.js';
+
+const getRandomInt = (min, max) => {
+  const lower = Math.ceil(Math.min(min, max));
+  const upper = Math.floor(Math.max(min, max));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+
+export {getRandomInt};
